+++ cloud/acp-spring-cloud-starter-common/src/main/java/pers/acp/springcloud/common/log/LogConstant.java
@@ -1,0 +1,13 @@
+package pers.acp.springcloud.common.log;
+
+/**
+ * @author zhangbin by 11/07/2018 17:36
+ * @since JDK1.8
+ */
+public interface LogConstant {
+
+    String INPUT = "lot_input";
+
+    String OUTPUT = "log_output";
+
+}